% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualCoxTable.R
\name{dualCoxTable}
\alias{dualCoxTable}
\title{Cox regression analysis for dual regulons}
\usage{
dualCoxTable(mbr, tns, duals = NULL, verbose = TRUE, excludeMid = FALSE)
}
\arguments{
\item{mbr}{an \linkS4class{MBR} object computed from
the same \linkS4class{TNI} objects that were used to make the
\linkS4class{TNS} objects.}

\item{tns}{a 'TNS' object with regulons used to compute the duals.}

\item{duals}{an optional vector containing duals to create the table. If NULL,
the regression is performed for all duals.}

\item{verbose}{a logical value. If TRUE, prints status of function while 
executing.}

\item{excludeMid}{if TRUE, doesn't use the samples that fall in mid section 
of the stratification for computations.}
}
\value{
A matrix containing the Cox regression results for all given duals
}
\description{
Returns a data.frame with Cox regression results for the interaction between 
regulons identified as duals.
}
\examples{
# load survival data
data(survival.data)

# load TNI-object
data(stni, package = "RTN")
stni <- upgradeTNI(stni)

# perform survival analysis
stns <- tnsPreprocess(stni, survival.data, keycovar = c('Grade','Age'),
                      time = 1, event = 2)
stns <- tnsGSEA2(stns, verbose=FALSE)

# create MBR-object using TF-TF duals
library(RTNduals)
tni <- tnsGet(stns, "TNI")
mbr <- tni2mbrPreprocess(tni, verbose = FALSE)
mbr <- mbrAssociation(mbr, pValueCutoff = 0.05)
results <- mbrGet(mbr, what="dualsCorrelation")

# create dual Cox regression
dualCox <- dualCoxTable(mbr, stns, verbose = FALSE)

}
\seealso{
\code{\link[RTNduals:tni2mbrPreprocess]{tni2mbrPreprocess}} for all 
plot parameters
}
