% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{dualSurvivalPanel,MBR-method}
\alias{dualSurvivalPanel,MBR-method}
\alias{dualSurvivalPanel}
\title{Panel plot of regulons in a inferred dual}
\usage{
\S4method{dualSurvivalPanel}{MBR}(mbr, tns1, tns2 = NULL, dual,
  attribs = NULL, endpoint = 60, path = NULL, nSections = 2,
  pal = "BrBG", excludeMid = FALSE, sectionsLegend = NULL,
  panelWidths = c(2, 3), png.res = 150)
}
\arguments{
\item{mbr}{an \linkS4class{MBR} object computed from
the same \linkS4class{TNI} objects that were used to make the
\linkS4class{TNS} objects.}

\item{tns1}{a `TNS` object with regulons used to compute the duals.}

\item{tns2}{another `TNS`  object computed with the other regulons used to 
compute the duals. It's optional if all duals are from the same 'TNI' object.}

\item{dual}{a character string with the name of a dual}

\item{attribs}{a numeric vector. Contains the columns of the survival 
data.frame which will be plotted as sample status in first and second plots.}

\item{endpoint}{a numeric value. It represents the cut-off point for the 
'time', if any.}

\item{path}{a string. The path to the directory where the plot will be saved.}

\item{nSections}{numeric value for the stratification of the sample. The 
larger the number, the more subdivisions will be created for the 
Kaplan-Meier and sample stratification analysis.}

\item{pal}{a palette name to be passed to \code{\link[RColorBrewer:brewer.pal]{brewer.pal}}.}

\item{excludeMid}{a logical value. if TRUE, doesn't use the samples that fall in mid section 
of the stratification for computations.}

\item{sectionsLegend}{a string vector. It must contain names for the sections
of the stratification on the interaction Kaplan-Meier. If NULL, the sections
are simply numbered in the legend.}

\item{panelWidths}{a numeric vector of length 2. It must contain a proportion
for the two panels in plots [1] and [2].}

\item{png.res}{a numeric value. It represents the resolution in ppi for the
png panel.}
}
\value{
A folder containing 5 pdf survival plots and one png survival panel
with all plots
}
\description{
Returns a folder with 5 plots that form a panel: 
[1] sample stratification based on activity of regulon 1, 
[2] sample stratification based on activity of regulon 2, 
[3] a scatter plot of samples's rank in both stratifications based on 
regulon activity
[4] Kaplan-Meier curves based on both regulon stratifications and interaction 
of regulons in dual,
[5] Cox regression for both regulons and the interaction between the activity
information for the dual. 
The function also returns a PNG panel with all plots aligned.
}
\examples{
# load survival data
data(survival.data)

# load TNI-object
data(stni, package = "RTN")
stni <- upgradeTNI(stni)

# perform survival analysis
stns <- tnsPreprocess(stni, survival.data, keycovar = c('Grade','Age'),
                      time = 1, event = 2)
stns <- tnsGSEA2(stns, verbose=FALSE)

# create MBR-object using TF-TF duals
library(RTNduals)
tni <- tnsGet(stns, "TNI")
mbr <- tni2mbrPreprocess(tni, tni, verbose = FALSE)
mbr <- mbrAssociation(mbr, prob = 0.75)
mbr <- mbrDuals(mbr)
duals <- mbrGet(mbr, what="dualRegulons")

# create panel plot

dualSurvivalPanel(mbr, stns, dual = duals[1], 
                  attribs = c("ER+", "ER-", "PR+", "PR-"))

}
\seealso{
\code{\link[RTNsurvival:tnsKM]{tnsKM}},
\code{\link[RTNsurvival:tnsCox]{tnsCox}}
}
