% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TNS-class}
\alias{TNS-class}
\alias{TNS}
\title{TNS: An S4 class for survival analysis using transcriptional 
networks inferred by the RTN package.}
\description{
TNS: An S4 class for survival analysis using transcriptional 
networks inferred by the RTN package.
}
\section{Slots}{

\describe{
\item{\code{TNI}}{a previously computed \linkS4class{TNI}-class object.}

\item{\code{survivalData}}{a data frame containing the survival data for all samples. 
Samples must be in the rows and the survival variables in the columns. 
Time of last update and event in last update (0 for alive, 1 for deceased).}

\item{\code{para}}{a list with the parameters used to compute the GSEA2 analysis.}

\item{\code{results}}{a list with results from TNS methods.}

\item{\code{status}}{a vector containing the processing status of the TNS object.}
}}

\section{Constructor}{


see \code{\link{tni2tnsPreprocess}} constructor.
}

