% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllSupplements.R
\name{tnsGet}
\alias{tnsGet}
\title{Get information from slots in a TNS object}
\usage{
tnsGet(object, what)
}
\arguments{
\item{object}{a TNS object}

\item{what}{a character vector specifying what should be retrieved from the
object. Options: "survivalData", "EScores", "TNI", "keycovar"}
}
\value{
A plot containing the 2-tailed GSEA analysis for a phenotype.
}
\description{
Get information from individual slots in a TNS object and 
any available results from a previous analysis.
}
\examples{
data(dt4rtn, package="RTN")
data(survival.data)

# retrieve the Enrichment Score results
rtni <- new("TNI", gexp=dt4rtn$gexp, 
transcriptionFactors=dt4rtn$tfs[c("FOXM1","E2F2","PTTG1")])
rtni <- tni.preprocess(rtni,gexpIDs=dt4rtn$gexpIDs, verbose=FALSE)
rtni<-tni.permutation(rtni, nPermutations=100, verbose=FALSE) #sets 'nPermutations'>=100
rtni<-tni.dpi.filter(rtni, verbose=FALSE)

rtns <- tnsPreprocess(rtni, survival.data, keycovar = c("Grade","Age"), 
time = 1, event = 2)
rtns <- tnsGSEA2(rtns, verbose=FALSE)
enrichmentScores <- tnsGet(rtns, "EScores")

}
