% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllSupplements.R
\name{tnsPlotGSEA2}
\alias{tnsPlotGSEA2}
\title{Plot 2-tailed GSEA for a sample from a TNS}
\usage{
tnsPlotGSEA2(object, aSample, regs = NULL, refsamp = NULL, log = FALSE,
  ntop = NULL, pValueCutoff = 0.05, pAdjustMethod = "BH",
  verbose = TRUE, ...)
}
\arguments{
\item{object}{a TNS object}

\item{aSample}{a string specifying a given sample number present in the 'survivalData' table.}

\item{regs}{an optional string vector specifying regulons to make the plot.}

\item{refsamp}{a character vector.}

\item{log}{a logical value. If TRUE, gexp values are transformed into log space.}

\item{ntop}{an optional integer value. The number of regulons for which the GSEA2 will be plotted.}

\item{pValueCutoff}{an integer. The p cutoff value for the analysis.}

\item{pAdjustMethod}{a character. Specifies the adjustment method for the pvalue.
See \code{\link[stats:p.adjust]{p.adjust}}}

\item{verbose}{a logical value. If TRUE, prints status of function while executing.}

\item{...}{parameters which will be passed to \code{\link[RTN:tna.plot.gsea2]{tna.plot.gsea2}},
such as ylimPanels, heightPanels, width, height, ylabPanels, xlab...}
}
\value{
A plot containing the 2-tailed GSEA analysis for a phenotype.
}
\description{
Makes a 2-tailed GSEA plot for a certain phenotype (sample)
present in a TNS. A wrapper of \code{\link[RTN:tna.plot.gsea2]{tna.plot.gsea2}}
}
\examples{
data(dt4rtn, package="RTN")
data(survival.data)

# compute regulons for 3 TFs using the RTN package
rtni <- new("TNI", gexp=dt4rtn$gexp, transcriptionFactors=dt4rtn$tfs[c("FOXM1","E2F2","PTTG1")])
rtni <- tni.preprocess(rtni,gexpIDs=dt4rtn$gexpIDs, verbose=FALSE)
rtni<-tni.permutation(rtni, nPermutations=100, verbose=FALSE) #sets 'nPermutations'>=1000
rtni<-tni.dpi.filter(rtni, verbose=FALSE)

rtns <- tnsPreprocess(rtni, survival.data, keycovar = c("Grade","Age"), time = 1, event = 2)
rtns <- tnsGSEA2(rtns, verbose=FALSE)
tnsPlotGSEA2(rtns, "MB-5115", regs = "FOXM1")

}
\seealso{
\code{\link[RTN:tna.plot.gsea2]{tna.plot.gsea2}} for all plot parameters
}

