\name{TNS.data}
\alias{survival.data}
\docType{data}

\title{A pre-processed dataset for demonstration purposes only.}

\description{
A minimum dataset used to demonstrate RTNsurvival main features.
}

\usage{data(survival.data)}

\format{
  \code{survival.data}{ A data.frame with a subset of samples in the Fletcher2013b package.}
}

\details{
The dataset consists of data.frame with survival and clinical variables used in the RTNsurvival vignettes. It should be regarded as a toy example for demonstration purposes only, despite being extracted, pre-processed and size-reduced from Fletcher et al. (2013) and Curtis et al. (2012).
}

\references{
Fletcher M.N.C. et al., \emph{Master regulators of FGFR2 signalling and breast cancer risk.} Nature Communications, 4:2464, 2013.

Curtis C. et al., \emph{The genomic and transcriptomic architecture of 2,000 breast tumours reveals novel subgroups.} Nature 486, 7403. 2012.
}

\value{
a data.frame.
}

\examples{
    data(survival.data)
}

\keyword{dataset}

