% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{tnsGet,TNS-method}
\alias{tnsGet,TNS-method}
\alias{tnsGet}
\title{Get information from slots in a TNS object}
\usage{
\S4method{tnsGet}{TNS}(object, what)
}
\arguments{
\item{object}{a TNS object}

\item{what}{a character vector specifying what should be retrieved from the
object. Options: 'survivalData', 'EScores', 'TNI', 'keycovar'}
}
\value{
A plot containing the 2-tailed GSEA analysis for a phenotype.
}
\description{
Get information from individual slots in a TNS object and 
any available results from a previous analysis.
}
\examples{
# load survival data
data(survival.data)

# load TNI-object
data(stni, package = "RTN")

stns <- tnsPreprocess(stni, survival.data, keycovar = c('Grade','Age'), 
time = 1, event = 2)
stns <- tnsGSEA2(stns, verbose=FALSE)
enrichmentScores <- tnsGet(stns, 'EScores')

}
