% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{tnsGSEA2,TNS-method}
\alias{tnsGSEA2,TNS-method}
\alias{tnsGSEA2}
\title{2-tailed Gene Set Enrichment Analysis on Transcriptional Networks.}
\usage{
\S4method{tnsGSEA2}{TNS}(tns, ...)
}
\arguments{
\item{tns}{A \linkS4class{TNS} class, which has been preprocessed}

\item{...}{Parameters passed to the \code{\link[RTN:tni.gsea2]{tni.gsea2}} 
function.}
}
\value{
A \linkS4class{TNS} class, with added Enrichment Scores.
}
\description{
Works as a wrapper for \code{\link[RTN:tni.gsea2]{tni.gsea2}}, performing a 
2-tailed GSEA analysis on a \linkS4class{TNI} class object and integrating 
the results into the \linkS4class{TNS} class object.
}
\examples{
# load survival data
data(survival.data)

# load TNI-object
data(stni, package = "RTN")

stns <- tnsPreprocess(stni, survival.data, keycovar = c('Grade','Age'), time = 1, event = 2)
stns <- tnsGSEA2(stns, verbose=FALSE)

}
\seealso{
\code{\link[RTN:tni.gsea2]{tni.gsea2}} for information on all 
parameters.
}
