% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllSupplements.R
\name{tnsPlotGSEA2}
\alias{tnsPlotGSEA2}
\title{Plot 2-tailed GSEA for a sample from a TNS}
\usage{
tnsPlotGSEA2(object, aSample, regs = NULL, refsamp = NULL, log = FALSE,
  ntop = NULL, pValueCutoff = 0.05, pAdjustMethod = "BH",
  verbose = TRUE, plotpdf = TRUE, ...)
}
\arguments{
\item{object}{a TNS object}

\item{aSample}{a string specifying a given sample number present in the 
'survivalData' table.}

\item{regs}{an optional string vector specifying regulons to make the plot.}

\item{refsamp}{a character vector.}

\item{log}{a logical value. If TRUE, gexp values are transformed into log 
space.}

\item{ntop}{an optional integer value. The number of regulons for which the 
GSEA2 will be plotted.}

\item{pValueCutoff}{an integer. The p cutoff value for the analysis.}

\item{pAdjustMethod}{a character. Specifies the adjustment method for the 
pvalue.
See \code{\link[stats:p.adjust]{p.adjust}}}

\item{verbose}{a logical value. If TRUE, prints status of function while 
executing.}

\item{plotpdf}{a single logical value.}

\item{...}{parameters which will be passed to 
\code{\link[RTN:tna.plot.gsea2]{tna.plot.gsea2}},
such as ylimPanels, heightPanels, width, height, ylabPanels, xlab...}
}
\value{
A plot containing the 2-tailed GSEA analysis for a phenotype.
}
\description{
Makes a 2-tailed GSEA plot for a certain phenotype (sample)
present in a TNS. A wrapper of \code{\link[RTN:tna.plot.gsea2]{tna.plot.gsea2}}
}
\examples{
# load survival data
data(survival.data)

# load TNI-object
data(stni, package = "RTN")

stns <- tnsPreprocess(stni, survival.data, keycovar = c('Grade','Age'), 
time = 1, event = 2)
stns <- tnsGSEA2(stns, verbose=FALSE)
tnsPlotGSEA2(stns, 'MB-5115', regs = 'FOXM1')

}
\seealso{
\code{\link[RTN:tna.plot.gsea2]{tna.plot.gsea2}} for all plot parameters
}
