% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllSupplements.R
\name{tnsStratification}
\alias{tnsStratification}
\title{Sample stratification for a TNS object}
\usage{
tnsStratification(tns, nSections = 2, setMid = FALSE)
}
\arguments{
\item{tns}{a \linkS4class{TNS} object, which must have passed GSEA2 analysis.}

\item{nSections}{A numeric value for the stratification of the sample. The 
larger the number, the more subdivisions will be created for the Kaplan-Meier 
analysis.}

\item{setMid}{a logical value. If it is TRUE, it assigns "zero" to the middle strata, 
centralizing the stratification on regulon activity.}
}
\value{
An updated \linkS4class{TNS} object.
}
\description{
Internal function, used for sample stratification.
}
\examples{
# see tnsKM method.

}
\seealso{
\code{\link{tnsKM}}
}
